Changed a bunch of debug and information printfs so that they have a better chance of working on systems where int is smaller than 32 bits also fixed the signed/unsigned type of some to match the type of the arguments.

+++ trunk/bacnet-stack/demo/handler/h_cov.c
@@ -385,8 +385,8 @@ void handler_cov_task(
             if (lifetime_seconds >= elapsed_seconds) {
                 COV_Subscriptions[index].lifetime -= elapsed_seconds;
 #if 0
-                fprintf(stderr, "COVtask: subscription[%d].lifetime=%d\n",
-                    index, COV_Subscriptions[index].lifetime);
+                fprintf(stderr, "COVtask: subscription[%d].lifetime=%lu\n",
+                    index, (unsigned long)COV_Subscriptions[index].lifetime);
 #endif
             } else {
                 COV_Subscriptions[index].lifetime = 0;
