+++ trunk/bacnet-stack/demo/handler/h_iam.c
@@ -52,7 +52,7 @@ void handler_i_am_add(
 #endif
     if (len != -1) {
 #if PRINT_ENABLED
-        fprintf(stderr, " from %u, MAC = %d.%d.%d.%d.%d.%d\n", device_id,
+        fprintf(stderr, " from %lu, MAC = %d.%d.%d.%d.%d.%d\n", (unsigned long)device_id,
             src->mac[0], src->mac[1], src->mac[2], src->mac[3], src->mac[4],
             src->mac[5]);
 #endif
