+++ trunk/bacnet-stack/demo/handler/h_ihave.c
@@ -45,11 +45,11 @@ void handler_i_have(
     len = ihave_decode_service_request(service_request, service_len, &data);
     if (len != -1) {
 #if PRINT_ENABLED
-        fprintf(stderr, "I-Have: %s %d from %s %u!\r\n",
+        fprintf(stderr, "I-Have: %s %lu from %s %lu!\r\n",
             bactext_object_type_name(data.object_id.type),
-            data.object_id.instance,
+            (unsigned long)data.object_id.instance,
             bactext_object_type_name(data.device_id.type),
-            data.device_id.instance);
+            (unsigned long)data.device_id.instance);
 #endif
     } else {
 #if PRINT_ENABLED
