+++ trunk/bacnet-stack/demo/object/bo.c
@@ -179,7 +179,7 @@ char *Binary_Output_Name(
     static char text_string[32] = "";   /* okay for single thread */
 
     if (object_instance < MAX_BINARY_OUTPUTS) {
-        sprintf(text_string, "BINARY OUTPUT %u", object_instance);
+        sprintf(text_string, "BINARY OUTPUT %lu", (unsigned long)object_instance);
         return text_string;
     }
